pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gaia-soil"
version = "0.1.0"
description = "A foundation model for soil microbiome understanding"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Gaia Project Contributors"},
]
keywords = ["soil", "microbiome", "foundation-model", "metagenomics", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.30",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "anndata>=0.10",
    "biom-format>=2.1",
    "requests>=2.31",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.7",
    "isort>=5.12",
    "flake8>=6.1",
    "mypy>=1.5",
]
viz = [
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "umap-learn>=0.5",
]
tracking = [
    "wandb>=0.15",
]
all = ["gaia-soil[dev,viz,tracking]"]

[project.urls]
Homepage = "https://github.com/your-org/gaia"
Documentation = "https://gaia-soil.readthedocs.io"
Repository = "https://github.com/your-org/gaia"
Issues = "https://github.com/your-org/gaia/issues"

[tool.setuptools.packages.find]
include = ["gaia*"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=gaia --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
